requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain_openai>=0.1.0
langchain_experimental>=0.0.50
pandas>=2.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0
streamlit>=1.28.0
arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.0